requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.4
python-dotenv==1.0.1
google-generativeai==0.7.2
pydantic==2.8.2
python-multipart==0.0.9
